openembedded-core: avoid absolute paths for general utils
On some distro the path of utils like sed, grep etc are different than the
buildhost system.
This was causing runtime issue while running libtoolize from the
libtool-nativesdk package.
This patch takes out the absolute paths of these utils, and put
responsibility on the end-user to set correct paths for these utilities to be
found out.

http://cgit.openembedded.org/openembedded-core/plain/meta/recipes-devtools/libtool/libtool/avoid_absolute_paths_for_general_utils.patch